fix(code-gen/java): Updated the codegen for java to generally use smaller types for unsigned integers (a byte is able to keep an unsigned integer value to 7 bits and not up to 4 bits)
feat(java): Added PlcBINT, PlcUBINT and PlcBREAL plc-value types

+++ plc4j/drivers/s7/src/test/java/S7IoTest.java
@@ -338,7 +338,7 @@ public class S7IoTest {
                     (short) 0
                 ),
                 false,
-                (short) 13
+                (byte) 13
             )
         );
         // To string
